+++ HostMachine/HostMachine/dlgfileexport.h
@@ -10,11 +10,11 @@ class DlgFileExport : public QDialog
     Q_OBJECT
 
 public:
-    DlgFileExport(float filesize, QWidget *parent = 0);
+    DlgFileExport(qint32 filesize, QWidget *parent = 0);
     ~DlgFileExport();
 
-    float Exportsize() const { return exportsize; }
-    float Startpos() const { return startpos; }
+    qint32 Exportsize() const { return exportsize; }
+    qint32 Startpos() const { return startpos; }
     QString ExportPath() const { return m_sExportPath; }
 
 private:
@@ -30,7 +30,7 @@ private:
     QLineEdit *m_lineEdit;
     QPushButton *m_btnOk, *m_btnCancel, *m_btnBrowser;
     QDoubleSpinBox *m_spinBox1, *m_spinBox2;
-    float m_filesize, startpos, exportsize;
+    qint32 m_filesize, startpos, exportsize;
     QString m_sExportPath;
 };
